refactor: split cli.py into modules by concern
cli.py was 1357 lines and held four unrelated jobs. It is now 872, under
pylint's limit, and each concern has its own module:

- apktool.py  driving apktool, working out and preparing the decompile
              directory, and restoring the original dex files
- smali.py    locating the main activity's smali, inserting the loadLibrary
              call, and editing the manifest
- signing.py  handing the rebuilt APK to uber-apk-signer
- paths.py    ROOT_DIR and FILE_DIR, which the gadget download and the signer
              download both need

cli.py keeps choosing and downloading the gadget, copying it into the APK
with its config and script, and the click command.

The tests follow the same split: test_apktool.py, test_smali.py and
test_signing.py alongside test_cli.py.

No behaviour change. Every definition was moved by line range rather than
retyped, and comparing the AST of each one against trunk shows all 39 are
byte-identical; only `p = Path(__file__)` disappears, folded into paths.py.
Running the injection against a synthetic APK on both revisions gives
identical output for all ten combinations of --js, --config,
--custom-gadget-name and --arch.
