+++ scripts/cli.py
@@ -1,158 +1,43 @@
-"""Frida Gadget Injector for Android APK.
+"""Command line entry point for injecting the Frida gadget into an APK.
 
-This script allows you to inject the Frida gadget library into an Android APK.
-It provides various functionalities including:
-- Decompiling the APK using apktool
-- Downloading the appropriate Frida gadget library based on the device architecture
-- Injecting the Frida gadget library into the APK
-- Modifying the AndroidManifest.xml to add necessary permissions
-- Recompiling the APK
-- Optionally signing the APK using uber-apk-signer
+The work is spread across the sibling modules:
+- apktool: decompiling, recompiling, and restoring the original dex files
+- smali: patching the main activity and the manifest
+- signing: handing the rebuilt APK to uber-apk-signer
+
+What is left here is choosing and downloading the gadget, copying it into the
+APK together with its config and script, and the click command tying it up.
 """
-import os
-import sys
+
 import atexit
+import json
 import shutil
 import subprocess
-import json
+import sys
 import tempfile
-import zipfile
-from shutil import which
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional
+
 import click
 from androguard.core.apk import APK
-from .apk_utils import get_main_activity
-from .logger import logger
-from .__version__ import __version__
-from .frida_github import FridaGithub
-from .uber_apk_signer_github import UberApkSignerGithub
+
 from . import INSTALLED_FRIDA_VERSION
-
-
-p = Path(__file__)
-ROOT_DIR = p.parent.resolve()
-FILE_DIR = ROOT_DIR.joinpath("files")
-
-APKTOOL = which("apktool")
-
-
-def get_decompiled_path(apk_path: Path) -> Path:
-    """Build the path apktool decompiles the APK into.
-
-    Args:
-        apk_path (Path): path of apk file
-
-    Returns:
-        Path: the decompile directory, next to the APK file
-    """
-    resolved = apk_path.resolve()
-    decompiled_path = resolved.with_suffix("")
-    if decompiled_path == resolved:
-        # The APK has no extension to strip, so keep the directory separate
-        decompiled_path = resolved.with_name(resolved.name + "_decompiled")
-    return decompiled_path
-
-
-def is_reusable_decompile_dir(path: Path) -> bool:
-    """Check whether a decompile directory can be safely removed.
-
-    Only empty directories and previous apktool outputs are reusable, so an
-    unrelated directory that happens to share the APK name is never deleted.
-
-    Args:
-        path (Path): path of the decompile directory
-
-    Returns:
-        bool: True if the directory can be removed
-    """
-    if not path.is_dir():
-        return False
-
-    entries = list(path.iterdir())
-    if not entries:
-        return True
-
-    # 'apktool.yml' is only written once decoding finishes, so the artifacts
-    # of an interrupted or failed decompile are accepted as well.
-    return any(
-        entry.name in ("apktool.yml", "AndroidManifest.xml", "original", "unknown")
-        or entry.name.startswith("smali")
-        for entry in entries
-    )
-
-
-def resolve_apktool(apktool_path: str = None) -> str:
-    """Work out the apktool command to run.
-
-    Args:
-        apktool_path (str): path or command given with --apktool-path
-
-    Returns:
-        str: the command to invoke apktool with
-
-    Raises:
-        FileNotFoundError: apktool is neither given nor on PATH
-    """
-    if not apktool_path:
-        if not APKTOOL:
-            raise FileNotFoundError(
-                "apktool not found. Please install apktool and add it to your PATH environment.\n"
-                "For macOS: brew install apktool\n"
-                "For Windows: Download from https://ibotpeaches.github.io/Apktool/install/\n"
-                "For Linux: sudo apt-get install apktool\n"
-                "After installation, you may need to restart your terminal."
-            )
-        return APKTOOL
-
-    apktool_parts = apktool_path.split()
-    apktool_binary = apktool_parts[-1]
-    if not Path(apktool_binary).exists():
-        logger.error("The specified apktool path does not exist: %s", apktool_binary)
-        sys.exit(-1)
-
-    if len(apktool_parts) > 1:
-        logger.info("Using custom apktool command: '%s'", apktool_path)
-    else:
-        logger.info("Using custom apktool path: '%s'", apktool_path)
-
-    return apktool_path
-
-
-def run_apktool(option: list, apk_path: str, apktool: str = None):
-    """Run apktool with option.
-
-    Args:
-        option (list|str): option of apktool
-        apk_path (str): path of apk file
-        apktool (str): command to invoke apktool with, defaults to the one on PATH
-    """
-    pipe = subprocess.PIPE
-    cmd = (apktool or APKTOOL).split() + option + [apk_path]
-    with subprocess.Popen(
-        cmd, stdin=pipe, stdout=sys.stdout, stderr=sys.stderr
-    ) as process:
-        process.communicate(b"\n")
-        if process.returncode != 0:
-            # Suggest only the options that are not in use yet. Removing from the
-            # list while iterating over it used to skip the second entry.
-            recommend_options = [
-                opt for opt in ("--no-res", "--use-aapt2") if opt not in option
-            ]
-
-            logger.error(
-                "It looks like you're having trouble with apktool.\n"
-                "Consider trying the '%s' options, or if you'd prefer more control,\n"
-                "you can manually specify apktool settings using "
-                "['--decompile-opts', '--recompile-opts', '--apktool-path'].",
-                recommend_options,
-            )
-
-            raise subprocess.CalledProcessError(
-                process.returncode, cmd, sys.stdout, sys.stderr
-            )
-        return True
+from .__version__ import __version__
+from .apk_utils import get_main_activity
+from .apktool import (
+    build_decompile_options,
+    get_decompiled_path,
+    prepare_decompile_dir,
+    resolve_apktool,
+    restore_original_dex,
+    run_apktool,
+)
+from .frida_github import FridaGithub
+from .logger import logger
+from .paths import FILE_DIR
+from .signing import sign_apk
+from .smali import insert_loadlibary, modify_manifest
 
 
 def download_gadget(
@@ -210,156 +95,6 @@ def download_gadget(
             )
 
     raise FileNotFoundError(f"'{file}' not found in the github releases")
-
-
-def download_signer():
-    """Download the Uber Apk Signer."""
-    signer_github = UberApkSignerGithub()
-    assets = signer_github.get_assets()
-    file = f"uber-apk-signer-{signer_github.signer_version}.jar"
-    signer_path = str(FILE_DIR.joinpath(file))
-    if os.path.exists(signer_path):
-        return signer_path
-
-    logger.debug("Downloading the %s file for signing", file)
-    return signer_github.download_signer_jar(assets, signer_path)
-
-
-def find_activity_smali(decompiled_path, main_activity):
-    """Locate the smali file of an activity across the smali_classes* trees.
-
-    Args:
-        decompiled_path (Path): decompiled path of apk file
-        main_activity (str): activity class name
-
-    Returns:
-        tuple: the smali file path, and the smali_classes<N> number it sits in
-               (None for the plain 'smali' directory)
-
-    Raises:
-        FileNotFoundError: the activity has no smali file
-    """
-    target_relative_path = main_activity.replace(".", os.sep) + ".smali"
-
-    for directory in sorted(decompiled_path.iterdir()):
-        if not directory.is_dir() or not directory.name.startswith("smali"):
-            continue
-
-        target_smali = directory.joinpath(target_relative_path)
-        if not target_smali.exists():
-            continue
-
-        class_number = None
-        if directory.name.startswith("smali_classes"):
-            class_number = int(directory.name.split("smali_classes")[1])
-        return target_smali, class_number
-
-    raise FileNotFoundError(
-        f"The smali file for '{main_activity}' was not found under {decompiled_path}."
-    )
-
-
-def insert_load_library_call(text: list, load_library_name: str) -> bool:
-    """Insert the loadLibrary call into the first usable entrypoint.
-
-    Args:
-        text (list): lines of the smali file, modified in place
-        load_library_name (str): name passed to System.loadLibrary
-
-    Returns:
-        bool: True when the call was inserted
-    """
-    if load_library_name.startswith("lib"):
-        load_library_name = load_library_name[3:]
-
-    for entrypoint in (" onCreate(", "<init>"):
-        for idx, line in enumerate(text):
-            if not line.strip().startswith(".method") or entrypoint not in line:
-                continue
-
-            # The call needs a register, which only a method declaring .locals has
-            if idx + 1 >= len(text):
-                continue
-
-            declaration = text[idx + 1]
-            if ".locals" not in declaration:
-                continue
-
-            # Increase the number of locals 0 to 1
-            text[idx + 1] = declaration.replace(".locals 0", ".locals 1")
-            text.insert(
-                idx + 2,
-                "    invoke-static {v0}, "
-                "Ljava/lang/System;->loadLibrary(Ljava/lang/String;)V",
-            )
-            text.insert(idx + 2, f'    const-string v0, "{load_library_name}"')
-            return True
-
-    return False
-
-
-def insert_loadlibary(decompiled_path, main_activity, load_library_name):
-    """Inject loadlibary code to main activity.
-
-    Args:
-        decompiled_path (str): decomplied path of apk file
-        main_activity (str): main activity of apk file
-        load_library_name (str): name of load library
-
-    Returns:
-        int: the smali_classes<N> number the patched activity sits in
-    """
-    logger.debug("Searching for the main activity in the smali files")
-    target_smali, target_smali_class_number = find_activity_smali(
-        decompiled_path, main_activity
-    )
-
-    logger.debug("Found the main activity at '%s'", str(target_smali))
-    text = target_smali.read_text(encoding="utf-8")
-    text = text.replace("invoke-virtual {v0, v1}, Ljava/lang/Runtime;->exit(I)V", "")
-    text = text.split("\n")
-
-    logger.debug("Locating the entrypoint method and injecting the loadLibrary code")
-    if not insert_load_library_call(text, load_library_name):
-        logger.error("Cannot find the appropriate position in the main activity.")
-        logger.error(
-            "Please report the issue at %s with the following information:",
-            "https://github.com/ksg97031/frida-gadget/issues",
-        )
-        logger.error("APK Name: <Your APK Name>")
-        logger.error("APK Version: <Your APK Version>")
-        logger.error("APKTOOL Version: <Your APKTOOL Version>")
-        sys.exit(-1)
-
-    # Replace the smali file with the new one
-    target_smali.write_text("\n".join(text), encoding="utf-8")
-    return target_smali_class_number
-
-def modify_manifest(decompiled_path):
-    """Modify manifest permissions.
-
-    Args:
-        decompiled_path (str): decomplied path of apk file
-    """
-    # Add internet permission
-    logger.debug("Checking internet permission and extractNativeLibs settings")
-    android_manifest = decompiled_path.joinpath("AndroidManifest.xml")
-    txt = android_manifest.read_text(encoding="utf-8")
-    pos = txt.index("</manifest>")
-    permission = "android.permission.INTERNET"
-
-    if permission not in txt:
-        logger.debug(
-            "Adding 'android.permission.INTERNET' permission to AndroidManifest.xml"
-        )
-        permissions_txt = f"<uses-permission android:name='{permission}'/>"
-        txt = txt[:pos] + permissions_txt + txt[pos:]
-
-    # Set extractNativeLibs to true
-    if ':extractNativeLibs="false"' in txt:
-        logger.debug('Editing the extractNativeLibs="true"')
-        txt = txt.replace(':extractNativeLibs="false"', ':extractNativeLibs="true"')
-    android_manifest.write_text(txt, encoding="utf-8")
 
 
 def detect_apk_architectures(decompiled_path):
@@ -704,169 +439,6 @@ def inject_gadget_into_apk(
     return insert_loadlibary(decompiled_path, main_activity, load_library_name)
 
 
-def stdin_is_interactive():
-    """Check whether a child process could prompt on this stdin.
-
-    Returns:
-        bool: True if stdin is attached to a terminal
-    """
-    try:
-        return sys.stdin is not None and sys.stdin.isatty()
-    except ValueError:  # stdin was already closed
-        return False
-
-
-def redact_passwords(cmd: list):
-    """Hide keystore passwords in a command line before it is reported.
-
-    Args:
-        cmd (list): command line of the signer
-
-    Returns:
-        list: copy of the command line without password values
-    """
-    redacted = list(cmd)
-    for idx, arg in enumerate(redacted[:-1]):
-        if arg in ("--ksPass", "--ksKeyPass"):
-            redacted[idx + 1] = "***"
-
-    return redacted
-
-
-def sign_apk(
-    apk_path: str,
-    ks: str = None,
-    ks_alias: str = None,
-    ks_key_pass: str = None,
-    ks_pass: str = None,
-):
-    """Run uber apk signer with option.
-
-    Args:
-        apk_path (str): path of apk file
-        ks (str): keystore file path
-        ks_alias (str): keystore alias
-        ks_key_pass (str): key password
-        ks_pass (str): keystore password
-    """
-    signer_path = download_signer()  # Download apk signer
-
-    cmd = ["java", "-jar", signer_path, "--apks", apk_path]
-
-    if ks:
-        cmd.append("--ks")
-        cmd.append(ks)
-    if ks_alias:
-        cmd.append("--ksAlias")
-        cmd.append(ks_alias)
-    if ks_key_pass:
-        cmd.append("--ksKeyPass")
-        cmd.append(ks_key_pass)
-    if ks_pass:
-        cmd.append("--ksPass")
-        cmd.append(ks_pass)
-
-    if ks_key_pass or ks_pass:
-        logger.warning(
-            "Keystore passwords given on the command line are readable by every "
-            "other user on this machine through the process list.\n"
-            "Omit the --ks-pass and --ks-key-pass options to let uber-apk-signer "
-            "prompt for them instead."
-        )
-
-    # uber-apk-signer prompts for the passwords that were not provided,
-    # which only works while it owns the terminal.
-    needs_prompt = bool(ks) and not (ks_pass and ks_key_pass)
-    interactive = needs_prompt and stdin_is_interactive()
-    if needs_prompt and not interactive:
-        logger.warning(
-            "There is no terminal attached, so uber-apk-signer cannot prompt for "
-            "the missing keystore password.\n"
-            "Run this from a terminal, or pass --ks-pass and --ks-key-pass."
-        )
-
-    stdio = None if interactive else subprocess.PIPE
-
-    with subprocess.Popen(
-        cmd, stdin=stdio, stdout=stdio, stderr=sys.stderr
-    ) as process:
-        stdout, _ = process.communicate(None if interactive else b"\n")
-        if process.returncode != 0:
-            logger.error("The APK signing process failed.")
-            raise subprocess.CalledProcessError(
-                process.returncode, redact_passwords(cmd), sys.stdout, sys.stderr
-            )
-
-        if stdout is None:  # the signer wrote directly to the terminal
-            return
-
-        output = stdout.decode()
-        print(output)
-        if "VERIFY" in output:
-            verify_message = output.split("VERIFY")[1]
-            if "file:" in verify_message:
-                apk_path = verify_message.split("file:")[1].split("\n")[0].strip()
-                logger.info("APK signing finished: %s", apk_path)
-
-
-def restore_original_dex(apk_path, recompiled_apk_path, modified_dex_number):
-    """Put the original dex files back into the recompiled APK.
-
-    apktool reassembles every dex file, which can change classes it was never
-    asked to touch. Only the dex holding the patched main activity has to come
-    from the rebuild; the rest are copied over from the original APK.
-
-    Args:
-        apk_path (Path): path of the original apk file
-        recompiled_apk_path (Path): path of the apk apktool produced
-        modified_dex_number (int): smali_classes<N> the main activity lives in
-    """
-    logger.debug(
-        "Copying original dex files (except modified one %s) to the recompiled APK",
-        modified_dex_number,
-    )
-
-    modified_dex_filename = (
-        f"classes{modified_dex_number}.dex"
-        if modified_dex_number and modified_dex_number > 1
-        else "classes.dex"
-    )
-
-    try:
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_dir_path = Path(temp_dir)
-            with zipfile.ZipFile(apk_path, "r") as original_apk:
-                original_apk.extractall(temp_dir_path)
-
-            # Built inside the temporary directory so a failure cannot leave a
-            # stray file behind the way the old NamedTemporaryFile did
-            staged_apk_path = temp_dir_path.joinpath("repacked.apk")
-            with zipfile.ZipFile(recompiled_apk_path, "r") as recompiled_apk, zipfile.ZipFile(
-                staged_apk_path, "w"
-            ) as new_apk:
-                for item in recompiled_apk.infolist():
-                    is_dex = item.filename.startswith("classes") and item.filename.endswith(
-                        ".dex"
-                    )
-                    if is_dex and item.filename != modified_dex_filename:
-                        dex_file = temp_dir_path.joinpath(item.filename)
-                        new_apk.write(str(dex_file), dex_file.name)
-                        continue
-
-                    if item.filename == modified_dex_filename:
-                        logger.debug("Copying %s from recompiled APK", item.filename)
-                    new_apk.writestr(item, recompiled_apk.read(item.filename))
-
-            shutil.move(str(staged_apk_path), str(recompiled_apk_path))
-            logger.info("Successfully replaced dex files in the recompiled APK")
-    except (OSError, zipfile.BadZipFile, zipfile.LargeZipFile, RuntimeError, KeyError) as e:
-        # Restoring the dex files is best effort: apktool has already produced a
-        # working APK, so a failure here is reported and the run carries on.
-        # RuntimeError covers the NotImplementedError zipfile raises for
-        # compression methods it cannot read, which must not take the run down.
-        logger.error("Failed to copy original dex files: %s", str(e))
-
-
 def detect_adb_arch():
     """Detect the architecture of the currently connected device via ADB.
 
@@ -1004,58 +576,6 @@ def validate_input_files(js: str, config: str):
     except json.JSONDecodeError:
         logger.error("The specified configuration file is not a valid JSON: %s", config)
         sys.exit(-1)
-
-
-def prepare_decompile_dir(decompiled_path: Path):
-    """Make sure the decompile target is an empty directory we may write to.
-
-    Args:
-        decompiled_path (Path): the decompile directory
-    """
-    if decompiled_path.exists():
-        if not is_reusable_decompile_dir(decompiled_path):
-            logger.error(
-                "The decompile target '%s' already exists and does not look like "
-                "an apktool output.\n"
-                "Remove or rename it, or use the --skip-decompile option "
-                "if it already contains the decompiled APK.",
-                decompiled_path,
-            )
-            sys.exit(-1)
-        shutil.rmtree(decompiled_path)
-
-    decompiled_path.mkdir(parents=True)
-
-
-def build_decompile_options(
-    decompiled_path: Path, force_manifest: bool, no_res: bool, decompile_opts: str
-):
-    """Assemble the apktool decode command line.
-
-    Args:
-        decompiled_path (Path): where apktool should write the decompiled APK
-        force_manifest (bool): whether --force-manifest was given
-        no_res (bool): whether --no-res was given
-        decompile_opts (str): extra options given with --decompile-opts
-
-    Returns:
-        tuple: the option list, and no_res updated from --decompile-opts
-    """
-    options = ["d", "-o", str(decompiled_path.resolve()), "--force"]
-    if force_manifest:
-        options += ["--force-manifest"]
-    if no_res:
-        options += ["--no-res"]
-
-    if decompile_opts:
-        if "--no-res" in decompile_opts:
-            if no_res:
-                # remove no-res option if it's already in the list
-                options.remove("--no-res")
-            no_res = True
-        options += decompile_opts.split()
-
-    return options, no_res
 
 
 def normalize_arch(arch: str, skip_decompile: bool) -> str:
